% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateGraph}
\alias{annotateGraph}
\title{Produce annotated gene graph where vertices have scores
attributed to genes.}
\usage{
annotateGraph(gH, gG, silent = FALSE, writeLog = TRUE)
}
\arguments{
\item{gH}{List of genes and their associated score (mutational frequncy of
mutSigCV value)}

\item{gG}{gene graph}

\item{silent}{Controls whether output to console should be suppressed. FALSE
by default.}

\item{writeLog}{Controls whether writing the result to the global logfile is
enabled. TRUE by default.}
}
\value{
AGG object with annotated vertices

  ## @examples ## \dontrun{
    annotateGraph(gH, gG)
  }
}
\description{
Produce annotated gene graph where vertices have scores
attributed to genes.
}
