% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{attachUUID}
\alias{attachUUID}
\title{Attach a UUID to an object}
\usage{
attachUUID(object, overwrite = TRUE)
}
\arguments{
\item{object}{any R object}

\item{overwrite}{flag for indicating whether a UUID should be overwritten or not}
}
\value{
NULL if no new UUID is attached, and the value of the attached UUID if a
new UUID was attached
}
\description{
\code{attachUUID} uses the uuid package to attach a $UUID attribute to
a given object.
}
\details{
The overwrite flag can be understood as follows:
- if the overwrite flag is TRUE, then whether or not the object had
a UUID previously or not, a new UUID attribute is attached
- if the overwrite flag is FALSE, then no new UUID attribute is attached
}
\examples{
\dontrun{
  object <- "c"
  attachUUID(object)
}
}
