% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSNV_CNA.R
\name{combineSNV_CNA}
\alias{combineSNV_CNA}
\title{Combine mutation data from CNA and SNV}
\usage{
combineSNV_CNA(fnameSNV = c(), fnameCNA = c(), fgX = "gX.rds",
  silent = FALSE, writeLog = TRUE)
}
\arguments{
\item{fnameSNV}{A vector of path of fully qualifed filenames of CNA rds files.}

\item{fgX}{The fully qualified filename/path of a gX output file, by default is gX.rds.}

\item{silent}{Boolean option for writing combining process information to console, FALSE by default.}

\item{writeLog}{Boolean option for log results, TRUE by default.}

\item{fNameCNA}{A vector of path of fully qualified file names of SNV rds files.}
}
\description{
\code{combineSNV_CNA} loads the vector of preprocessed dataset from filtered CNA or SNV files,
 creates hash keys corresponding to the input file type, combine the corresponding mutation count
 to each gene and save to a gX format file as output file.
}
\details{
Details.
}
\section{gX files}{

 gX rds files contain data of SNV and CNA combined for each gene as one object, sorted per
 gene.
}

\examples{
\dontrun{combine(fNameSNV="SNV.rds", fnameCNA="CNA.rds", fgX="combined_data.rds", silent=TRUE, writeLog=FALSE)}

}
