% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalData.R
\docType{data}
\name{devCNA.txt}
\alias{devCNA.txt}
\title{CNA data.}
\format{A tab separated text file with 16 rows and 582 columns, most of
  which represent individual cancer tissue samples:
\describe{
  \item{Gene Symbol}{HGNC gene symbol}
  \item{Locus ID}{Entrez Gene ID}
  \item{Cytoband}{Chromosome, arm and band information}
  \item{<Sample Barcode>...}{497 tissue sample barcodes}
  ...
}}
\source{
\url{https://tcga-data.nci.nih.gov/docs/publications/ov_2011/}
}
\description{
A file containing TCAG OV carcinoma CNAs for 16
genes.
}
\examples{
\dontrun{
system.file("extdata", "devCNA.txt", package="rete")
fPath <- system.file("extdata", "devCNA.txt", package="rete")
myCNA <- read.delim(fPath, stringsAsFactors = FALSE)
}
}
