% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalData.R
\docType{data}
\name{devPPI.txt}
\alias{devPPI.txt}
\title{PPI data.}
\format{A MITAB formatted text file with 75 rows and 54 columns:
\describe{
  \item{uidA}{UniProt ID of interactor A}
  \item{uidB}{UniProt ID of interactor B}
  ...
}}
\source{
\url{http://irefindex.org/}
}
\description{
A file containing protein-protein interactions for 16
genes, subset from iRefIndex version 14 data.
}
\examples{
\dontrun{
system.file("extdata", "devPPI.txt", package="rete")
fPath <- system.file("extdata", "devPPI.txt", package="rete")
myMITAB <- read.delim(fPath, stringsAsFactors = FALSE)
}
}
