% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalData.R
\docType{data}
\name{devSNV.maf}
\alias{devSNV.maf}
\title{SNV data.}
\format{A MAF formatted file with 15 rows and 32 columns:
\describe{
  \item{HUGO_Symbol}{HGNC gene symbol}
  \item{ENTREZ_GENE_ID}{ID}
  ...
}}
\source{
\url{https://tcga-data.nci.nih.gov/docs/publications/ov_2011/}
}
\description{
A MAF formattted file containing TCAG OV carcinoma SNVs for 16
genes.
}
\examples{
\dontrun{
system.file("extdata", "devSNV.maf", package="rete")
fPath <- system.file("extdata", "devSNV.maf", package="rete")
myMAF <- read.delim(fPath, skip = 1, stringsAsFactors = FALSE)
}
}
