% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{extractAttributes}
\alias{extractAttributes}
\title{Extract attributes from object}
\usage{
extractAttributes(object)
}
\arguments{
\item{object}{any R object}
}
\value{
NULL if no attributes are extracted, text of attributes if attributes exist
}
\description{
\code{extractAttributes} returns all of the attributes that are attached
to an object, formatted for logging
}
\details{
Format of the extracted attributes:
attribute | <attrName1> | <attrValue1>
attribute | <attrName2> | <attrValue2>
attribute | <attrName3> | <attrValue3>
}
\examples{
\dontrun{
  object <- "c"
  attr(object, "UUID") <- uuid::UUIDgenerate()
  extractAttributes(object)
}
}

