% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMap.R
\name{fastMap}
\alias{fastMap}
\title{map some ID type to another.}
\usage{
fastMap(ID, type = "UniProt")
}
\arguments{
\item{ID}{<description>.}

\item{type}{<description>.}
}
\description{
\code{fastMap} placeholder function
}
