% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMapUpdate.R
\name{fastMapUpdate}
\alias{fastMapUpdate}
\title{Modify a fastMap hash table.}
\usage{
fastMapUpdate(hashTable, key, value)
}
\arguments{
\item{hashTable}{The fastMap hash table.}

\item{key}{The key to update.}

\item{value}{The HGNC symbol to associate with \code{key}.}
}
\description{
\code{fastMapUpdate} Update, add, or delete entries in the fastMap
hash table. Adding and updating are the same; if a key does/doesn't
exist, will set the key's value to \code{value}. If a lookup is made
for a non-existant key, NULL will be returned; setting a key's value to
NULL is essentially deleting. Before inserting into the hash table, convert
all letters in \code{value} to upper case.
}
\details{
TODO: Vectorize.
}
\examples{
\dontrun{
fastMapUpdate(fastMapUniprot, "P01023", "A2M")
fastMapUpdate(fastMapENSP, "ENSP00000245105", "a2m-as1")
}
}
\seealso{
\code{\link{fastMapGenerate}} on how the hash table is structured.
\code{\link{fastMapSanity}} on acceptable keys and values.

Other fastMap functions: \code{\link{fastMapGenerate}},
  \code{\link{fastMapSanity}}, \code{\link{fastMap}}
}
