% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.lowExpresionGenes.R
\name{filter.lowExpressionGenes}
\alias{filter.lowExpressionGenes}
\title{Filter low expression genes}
\usage{
filter.lowExpressionGenes(exprData, rCNA, rSNV, dOut, whitelist, rT = 3,
  pT = 0.7, silent = FALSE, noLog = FALSE)
}
\arguments{
\item{exprData}{The normalized Firehose RNAseq files to scan for expression levels}

\item{rCNA}{The rCNA RDS files to filter out lower expression genes}

\item{dOut}{The local path to a directory to output the filtered files}

\item{whitelist}{The name of the file that contains genes to preserve}

\item{rT}{The minimum number of reads for genes to keep.  Default: 3}

\item{pT}{The percentage of samples that must have at least rT reads.  Default: 0.7}

\item{silent}{Whether or not to write progress information to console, default: FALSE}

\item{noLog}{Whether or not to log results, default: FALSE}

\item{rCNA}{The rSNV text files to filter out lower expression genes}
}
\value{
The list of filtered genes
}
\description{
This filter removes genes from rCNA and rSNV files that are not highly
expressed in the provided RNAseq files.
}
