% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSub.R
\name{findSub}
\alias{findSub}
\title{Find Highly Connected Subnetworks after Eliminating 'Weak' Edges,
and calculating aggregate 'heat' scores for these subnetworks}
\usage{
findSub(method = "Leis", EGG, minOrd, noLog = FALSE, silent = FALSE)
}
\arguments{
\item{method}{method of edge removal. Currently only "Leis" (Leiseron et al. 2015) method of edge
removal available. See details for more information}

\item{EGG}{an igraph object of format AGG with score annotations}

\item{minOrd}{minimum number of vertices per generated subgraph}

\item{noLog}{if FALSE, save process to an .Rdata log file}

\item{silent}{if FALSE, print process to console in real time}
}
\value{
A list of igraph objects, with "Influence" edge attributes,
"Gene Score" vertex attributes, "EGGversion" graph Attribute,
"Aggregate Heat score graph attribute
}
\description{
Find Highly Connected Subnetworks after Eliminating 'Weak' Edges,
and calculating aggregate 'heat' scores for these subnetworks
}
\details{
Depending on method used, edges are removed from EGG based upon the assigned 'delta' attribute
and their 'Influence' scores(see next paragraph). In the "Leis" method (Leiseron et al. 2015), edges with
Influence scores below the value of 'delta' are eliminated. Then, highly connected components of
the EGG graph made of the remaining edges are found using igraph::components(),
and subgraphs generated for these components. The vertices in these subgraphs come with the Gene Scores
(score can differ depending on method), which for each subgraph's vertices are summed to find an
aggregate heat score for each individual graph. E.g., if a subgraph's vertices had scores of 5,8,11
the aggregate heat score would be 24. The aggregate heat score is then assigned as an attribute
aggHeatScore. Subgraphs are stored in a list of subgraphs sN, which is returned. sN is ordered from
highest aggHeatScore to lowest aggHeatScore of the subgraph.
All subgraphs in sN contain same attributes as EGG and have aggHeatScore (aggregate heat score)
added as an attribute.

EGG must have the following: vertices: genes with HGNC symbol
and Gene Score ('heat', see documentation for score function for more details) as the minimal vertex
attributes; edges: directed edges from 1 gene to the next, derived from protein protein interaction data.
Edges' 'from' and 'to' columns (edge vertices) named with HGNC symbols. Minimum attributes for edges
include 'Influence', a score weighting edges depending on weighting method (for our purposes this
will likely ). EGG must also come with a delta attribute (integer or of type double)
(assigned threshold for edge removal regardless of method used in generating edge scores) and an
EGGversion, a graph attribute with a character vector of length 1. Unless further changes are made,
EGGversion must be "1.0".
}
