% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{findUUID}
\alias{findUUID}
\title{Finds specific UUID in logs}
\usage{
findUUID(uuid, uuidPath = getwd())
}
\arguments{
\item{uuid}{a universally unique identifier}

\item{uuidPath}{the path in which to search for the logfile, if given. Otherwise, uses getwd().}
}
\value{
NULL if uuid is not found in logs. Otherwise, returns complete event/comment blocks.
}
\description{
\code{findUUID} returns all mentions of a specific UUID in the logs, with all of the events
and comments attached to that specific UUID.
}
\examples{
\dontrun{

}
}
