% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{getProvenance}
\alias{getProvenance}
\title{Reconstructs provenance of object}
\usage{
getProvenance(uuid, uuidPath = getwd())
}
\arguments{
\item{uuid}{the UUID of the object of interest}

\item{uuidPath}{the path in which to search for logFile}
}
\value{

}
\description{
\code{getProvenance} reconstructs the provenance of an object back via its intermediaries
to all input files that contributed to it
}
\details{
Format of the output
}
\examples{
\dontrun{

}
}
