% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{getUUID}
\alias{getUUID}
\title{get a UUID to attach to an object}
\usage{
getUUID(objectName, overwrite = FALSE)
}
\arguments{
\item{objectName}{char. Name of an R object}

\item{overwrite}{logical. Flag to control whether an existing UUID should be
overwritten. Default FALSE.}
}
\value{
the value of the original UUID attribute, or a new UUID
}
\description{
\code{getUUID} uses the uuid package to prepare a UUID suitable to
               be attached to an object identified by objectName.
}
\details{
The overwrite flag can be understood as follows:
- if the object does not have a $UUID attribute, a UUID is returned
- if the object has a $UUID attribute and "overwrite" is TRUE
  a new UUID is returned.
- if the object has a $UUID attribute and "overwrite" is FALSE, the
  old UUID is returned.

Note: no check is done whether the original $UUID attribute
  is a valid UUID.
}
\examples{
\dontrun{
  tmp <- "c"
  getUUID("tmp")
}
}
