% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFilterHypermutators.R
\name{importFilterHypermutators}
\alias{importFilterHypermutators}
\title{Filters out hypermutators}
\usage{
importFilterHypermutators(rSNVFileIn = c(), rCNAFileIn = c(),
  dOut = getwd(), xS = 400, silent = FALSE, writeLog = TRUE,
  writeDetailedLog = TRUE)
}
\arguments{
\item{rSNVFileIn}{A vector of local file names of rSNV}

\item{rCNAFileIn}{A vector of local file names of rCNA}

\item{dOut}{Directory to store output, defaults to getwd().}

\item{xS}{Mutation threshold. 400 by default.}

\item{silent}{Controls whether output to console should be suppressed. FALSE
by default.}

\item{writeLog}{Controls whether writing the result to the global logfile is
enabled. TRUE by default.}

\item{writeDetailedLog}{Flag for extra details about log. TRUE by default.}
}
\description{
\code{importFilterHypermutators} identifies the number of SNV and CNA
mutations for each sample and removes the samples from rSNV and rCNA
datasets. Default threshold for removal is 400 mutations per sample.
}
\examples{
\dontrun{
    importFilterHypermutators(rSNVFileIn, rCNAFileIn, dOut, xS)
}

}
