% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNet.STRING.R
\name{importNet.STRING}
\alias{importNet.STRING}
\title{Import network data from a STRING database file.}
\usage{
importNet.STRING(fName, net = "combined_score", cutoffType = "xN", val,
  taxID = "9606", silent = FALSE, noLog = FALSE)
}
\arguments{
\item{<fName>}{<description>.}

\item{<net>}{<description>.}

\item{<cutoffType>}{<description>.}

\item{<val>}{<description>.}

\item{<taxID>}{<description>.}

\item{<silent>}{<description>.}

\item{<noLog>}{<description>.}
}
\value{
<description>.
}
\description{
\code{importNet.STRING} description.
}
\details{
Details.
}
\section{<title>}{
 Additional explanation.
}

