% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSNV.TCGA.R
\name{importSNV.TCGA}
\alias{importSNV.TCGA}
\title{Import MAF files from source and converts them to rSNV files.}
\usage{
importSNV.TCGA(fMAF, rSNV, na.rm = FALSE, silent = FALSE, writeLog = TRUE)
}
\arguments{
\item{fMAF}{vector of MAF file names.}

\item{rSNV}{rSNV file to merged new MAF data with.}

\item{na.rm}{Remove all rows with missing or NA values and drop them from data frame.}

\item{silent}{Controls whether output to console should be suppressed. FALSE
by default.}

\item{writeLog}{Controls whether writing the result to the global logfile is
enabled. TRUE by default.}
}
\value{
N/A. Function is invoked for its side-effect of saving an rSNV object in RDS format to the requested filename.
}
\description{
\code{importSNV.TCGA} Imports data from one or more files of MAF format, creates rSNV or adds to
existing rSNV file, returns rSNV file.
}
\section{Validations}{
 validations sections has 3 functions - function that checks MAF file extension,
function that checks MAF file header, function that checks rSNV file header.
}

\section{Read MAF file}{
 maf file is read by extracting require columns from maf file followed
by checking if every item in the class, type and chr columns are from a known and required list of values.
These columns as a data table are then added to an rSNV file.
}

\section{Write Log}{
 log section writes file processing progress along with other errors
that may occur.
}

