% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{logEvent}
\alias{logEvent}
\title{Main function for writing logs}
\usage{
logEvent(eventTitle = NULL, eventCall = NULL, input = list(),
  output = list(), fPath = getwd())
}
\arguments{
\item{eventTitle}{the title of the event, from a predetermined categorization of events}

\item{eventCall}{the function call in which the event occurred}

\item{input}{the list of input objects - given a name that is both an object and filename, defaults to
treating it as an object}

\item{output}{the list of output objects - given a name that is both an object and filename, defaults to
treating it as an object}

\item{fPath}{the file path in which to log the event, defaults to getwd()}
}
\value{
N/A. message is appended to the logFile with \code{\link{logMessage}}.
}
\description{
\code{logEvent} generates a log of an event with:
- the event title
- the event call
- attributes of the input object
- attributes of the output object
- date and time of the event
- function version (TODO)
- file hashes of input (TODO)
}
\details{
Format of the logging is as follows:
event | input | attribute | <inputAttrName1> | <inputAttrValue1>
event | output | attribute | <outputAttrName1> | <outputAttrValue1>
event | call | <eventCall>
comment | title | <eventTitle>
comment | dateTime | <eventDateTime>
comment | functionVersion | <functionVersion>
}
\examples{
\dontrun{
    logEvent(eventTitle = "Test event", eventCall = "function1(test = \\"test\\")")
}
}
