% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCoord.R
\name{mapCoord}
\alias{mapCoord}
\title{Populate a GeneData list with functions for mapping coordinates from chromosome to gene}
\usage{
mapCoord(geneData, outFile = "geneData.rds", dropExonBoundaries = TRUE,
  silent = FALSE, writeLog = TRUE)
}
\arguments{
\item{geneData}{the GeneData formatted resource list with at least strand,
start, end, exonStart, exonEnd elements populated}

\item{outFile}{File where the populated GeneData list is stored
(default: "geneData.rds")}

\item{dropExonBoundaries}{logical. Populated GeneData list has exonStart
and exonEnd elements dropped iff \code{TRUE}. (default: \code{TRUE})}

\item{silent}{logical. Console is suppressed iff silent is \code{TRUE}.
(default: \code{FALSE})}

\item{writeLog}{logical. Log file is written iff \code{TRUE}.
(default: \code{TRUE})}
}
\description{
\code{mapCoord.R} populates map elements of a GeneData list with a
corresponding function closure that maps chromosomal coordinates to
gene coordinates.
The populated GeneData list is saved to an .rds file.
}
\section{Mapping function closure}{

The mapping function that \code{mapCoord.R} creates and populates a GeneData
list which takes a chromosomal coordinate as input and returns the
corresponding gene coordinate if the chromosomal coordinate occurs in an
exon of the gene.
If the input is in an intron, the function returns 0.
If the input is out-of-bounds of the gene or invalid, the function
returns \code{NULL}.
}

