% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply two numbers.}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The product of \code{x} and \code{y}. It will be cast to the most
  general type of its arguments.
}
\description{
\code{multiply} returns the product of its two arguments.
}
\details{
This is an inconvenience function that achieves exactly the same as the
\code{"*"} operator, which it uses internally, but with more characters to
type.
}
\section{QA warning}{
 Actually using this function instead of "*" will get you
  in trouble with QA.
}
\examples{
multiply(1, 1)
multiply(1i, 1i)
multiply(13, 3)
}
\seealso{
\code{\link[base]{prod}} for products, \code{\link[base]{*}} for the
  arithmetic operator.
}

