% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{rete_content_hash}
\alias{rete_content_hash}
\title{Content hash}
\usage{
rete_content_hash(graph)
}
\arguments{
\item{graph}{A \code{rete_graph}.}
}
\value{
The file's blake3-16 content hash as a hex string.
}
\description{
Content hash
}
