% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{rete_info}
\alias{rete_info}
\title{Graph size and header counts}
\usage{
rete_info(graph)
}
\arguments{
\item{graph}{A \code{rete_graph}.}
}
\value{
A list with \code{quads}, \code{terms}, \code{pyramidLevels}, and \code{namedGraphs}.
}
\description{
Graph size and header counts
}
