% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMapGenerate.R
\name{fastMapGenerate}
\alias{fastMapGenerate}
\title{Generate the fastMap hash tables.}
\usage{
fastMapGenerate(fName, HGNCSymbolColName, unmappedColName, type,
  saveHashTable = TRUE, outputName)
}
\arguments{
\item{fName}{The path to the HGNC gene symbol data set.}

\item{HGNCSymbolColName}{The column name of the HGNC symbol.}

\item{unmappedColName}{The column name the unmapped symbol.}

\item{type}{The type of the gene database of unmappedColName.}

\item{saveHashTable}{Boolean whether to save the hash table as an rds.}

\item{outputName}{The path to save the RDS.}
}
\description{
\code{fastMapGenerate} will parse through an HGNC file and create a hash
table. Keys are the IDs found in \code{unmappedColName} and associated values
are IDs found in \code{HGNCSymbolColName}. The type attribute of the hash
table will be set to \code{type}. Hash table can be saved and can be loaded
in future sessions. If there are conflicting keys or a tab delimited key, keep
the first instance of the key or only store the first key respectively.
}
\examples{
\dontrun{
fastMapGenerate("hgnc_complete_set.txt", "symbol", "uniprot_ids",
                type = "UniProt", outputName = "fastMapUniProt.rds")
}
}
\seealso{
\code{\link{fastMapSanity}} on acceptable keys and values.

Other fastMap functions: \code{\link{fastMapSanity}},
  \code{\link{fastMapUpdate}}, \code{\link{fastMap}}
}
