% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMapSanity.R
\name{fastMapSanity}
\alias{fastMapSanity}
\title{Check if given input is a valid key or valid value}
\usage{
fastMapSanity(toCheck)
}
\arguments{
\item{toCheck}{Given input to check.}
}
\description{
\code{fastMapSanity} Will check if \code{toCheck} contains an illegal
character (Allowed characters: alphanumeric, -, _).
}
\examples{
fastMapSanity("ENSP00000245105")
fastMapSanity("A2M")
}
\seealso{
Other fastMap functions: \code{\link{fastMapGenerate}},
  \code{\link{fastMapUpdate}}, \code{\link{fastMap}}
}
