% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.inconsistentCNA.R
\name{filter.inconsistentCNA}
\alias{filter.inconsistentCNA}
\title{Filter inconsistent CNA entries}
\usage{
filter.inconsistentCNA(rCNAfile, fNames, dOut, filterFile, iT = 0.75)
}
\arguments{
\item{fNames}{A vector of filenames to filter}

\item{dOut}{The local path to a directory to output the filtered files}

\item{filterFile}{The filename to output the list of genes removed}

\item{iT}{The inconsistency threshold to use.  Default: 0.75}

\item{rCNA}{The rCNA RDS file that contains the CNAs to consider}
}
\value{
The list of filtered genes
}
\description{
This filter removes genes from rCNA files that have inconsistent signs.
Raw rCNA files can contain genes whose copy numbers are consistently
increased or decreased relative to normal expression.  This filter finds
genes whose expression is not consistently increased or decreased, then
removes them from the dataset.
}
