% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{findUUID}
\alias{findUUID}
\title{Finds a specific UUID in logged events}
\usage{
findUUID(uuid, logDir, ext = "\\\\.log$", recursive = TRUE)
}
\arguments{
\item{uuid}{a universally unique identifier (UUID). Checked for valid format.}

\item{logDir}{the directory path in which to search the logfiles.
Defaults to the path component of getOption("rete.logfile").}

\item{ext}{file extension regex pattern. Defaults to "\\.log$".}

\item{recursive}{Whether to descend into subdirectories. Defaults
to TRUE.}
}
\value{
character() if uuid is not found in the log files.
        Otherwise, a vector with one complete event block per element, which
        has the filename that conteined the event attaches as a names()
        attribute.
}
\description{
\code{findUUID} returns the event blocks that contains mentions of a
requested UUID from the log files found in a given directory.
}
\examples{
\dontrun{
NULL
}
}
