% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFilterHypermutators.R
\name{importFilterHypermutators}
\alias{importFilterHypermutators}
\title{Filters out hypermutators}
\usage{
importFilterHypermutators(fNames = c(), dOut = getwd(), xS = 400,
  silent = FALSE, writeLog = TRUE)
}
\arguments{
\item{fNames}{A vector of local file names of rMUT and/or rCNA files}

\item{dOut}{Directory to store output, defaults to getwd().}

\item{xS}{Mutation threshold. 400 by default.}

\item{silent}{Controls whether output to console should be suppressed. FALSE
by default.}

\item{writeLog}{Controls whether writing the result to the global logfile is
enabled. TRUE by default.}
}
\description{
\code{importFilterHypermutators} identifies the number of SNV and CNA
mutations for each sample and removes the samples from rMUT and rCNA
datasets. Default threshold for removal is 400 mutations per sample.
}
\examples{
\dontrun{
    importFilterHypermutators(fNames, dOut, xS)
}

}
