% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{logEvent}
\alias{logEvent}
\title{Formats an event description to be sent for attaching to the log file.}
\usage{
logEvent(eventTitle, eventCall, input = character(), notes = character(),
  output = character())
}
\arguments{
\item{eventTitle}{the title of the event, from a
predetermined categorization of events}

\item{eventCall}{the function call in which the event occurred}

\item{input}{vector of object names with an "input" role in the calling
function's workflow}

\item{notes}{a vector of strings comprising text to be
incorporated into the event description.}

\item{output}{vector of object names  with an "output" role in the calling
function's workflow}
}
\value{
N/A A message is appended to the log file
            via \code{\link{logMessage}}.
}
\description{
\code{logEvent} generates an event message with:
- the event title
- the event call
- attributes of the input object
- attributes of the output object
- date and time of the event
- function version (TODO)
- file hashes of input (TODO)
}
\details{
Format of the event description is as follows:
event | title  | <eventTitle>
event | time   | <eventDateTime>
event | call   | <eventCall>
(ToDo: event | functionVersion | <functionVersion>)
event | input  | attribute | <inputAttrName1> | <inputAttrValue1>
event | output | attribute | <outputAttrName1> | <outputAttrValue1>
event | end

The event message is terminated by an "end event" marker, and a blank
line, and it is handed off to logMessage() to append it to the
log file referenced in the global variable options("rete.logfile")
}
\examples{
\dontrun{
    logEvent(eventTitle = "Test event",
             eventCall = "function1(test = \\"test\\")")
}
}
