% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTools.R
\name{logMessage}
\alias{logMessage}
\title{Write a message to the current log file.}
\usage{
logMessage(msg)
}
\arguments{
\item{msg}{a character object or vector of character objects.}
}
\value{
N/A. This function is invoked for its side-effect to appended text to
  the current logfile.
}
\description{
\code{logMessage} uses cat() to append a message to the current log
  file. The file name is taken from the rete.logfile global option.
}
\details{
The function will stop() if message is not of mode, type and class character.
On windows systems, cat() replaces \\n linebreaks with \\r\\n. Therefore
logMessage() converts all linebreaks internally to \\n before handing them to
cat().
}
\examples{
\dontrun{
  msg <- c("note > ", Sys.Date(), "Re-run analysis.")
  logMessage(msg)
}
}
